Safari-609.1.11.1/LayoutTests/webgl/max-active-contexts-oldest-context-lost-expected.txt

CONSOLE MESSAGE: line 9: There are too many active WebGL contexts on this page, the oldest context will be lost.
PASS if the first context was lost due to creating too many WebGL contexts.
